requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
SQLAlchemy>=2.0.32
Werkzeug==3.0.1
gunicorn>=21.0.0
